requirements.txt
Django==4.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-environ==0.11.2
django-cors-headers==4.3.1
six==1.16.0
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
django-tinymce==3.7.1
